Refactor ExtractorMerger to break a circular import
