+++ src/datumaro/plugins/data_formats/datumaro/importer.py
@@ -5,9 +5,9 @@
 import os.path as osp
 from typing import Dict, List, Optional, Type
 
-from datumaro.components.dataset_base import ExtractorMerger
 from datumaro.components.format_detection import FormatDetectionConfidence, FormatDetectionContext
 from datumaro.components.importer import Importer
+from datumaro.components.merge.extractor_merger import ExtractorMerger
 from datumaro.util import parse_json
 
 from .format import DatumaroPath
