+++ src/datumaro/plugins/data_formats/yolo/importer.py
@@ -7,10 +7,10 @@ from collections import defaultdict
 from io import TextIOWrapper
 from typing import Any, Dict, List, Optional, Type
 
-from datumaro.components.dataset_base import ExtractorMerger
 from datumaro.components.errors import DatasetImportError
 from datumaro.components.format_detection import FormatDetectionConfidence, FormatDetectionContext
 from datumaro.components.importer import Importer
+from datumaro.components.merge.extractor_merger import ExtractorMerger
 from datumaro.util.os_util import extract_subset_name_from_parent
 
 from .format import YoloFormatType, YoloLoosePath, YoloPath, YoloUltralyticsPath
